Create functionality to track enemy units that go out of vision and used for scout.

+++ javascript/systems/track-enemy-system.js
@@ -2,15 +2,22 @@
 "use strict"
 
 const { createSystem } = require("@node-sc2/core");
+const { Alliance } = require("@node-sc2/core/constants/enums");
+const { distance } = require("@node-sc2/core/utils/geometry/point");
 const enemyTrackingService = require("./enemy-tracking/enemy-tracking-service");
 
 module.exports = createSystem({
   name: 'TrackEnemySystem',
   type: 'agent',
+  async onStep({resources}) {
+    const { units } = resources.get();
+    enemyTrackingService.clearOutdatedMappedUnits(resources);
+    enemyTrackingService.addUnmappedUnit(units);
+  },
   async onEnemyFirstSeen({}, seenEnemyUnit) {
     enemyTrackingService.addEnemyUnit(seenEnemyUnit);
   },
   async onUnitDestroyed({}, unitDestroyed) {
-    enemyTrackingService.removeEnemyUnit(unitDestroyed);
+    if (unitDestroyed.alliance === Alliance.ENEMY) enemyTrackingService.removeEnemyUnit(unitDestroyed);
   }
 });
